Replace t2rust with embedding bincode-serialized target descriptions.
