+++ Cargo.toml
@@ -2,7 +2,6 @@
 members = [
     "probe-rs",
     "probe-rs-target",
-    "probe-rs-t2rust",
     "probe-rs-cli-util",
     "cli",
     "gdb-server",
