+++ doc/release.md
@@ -5,7 +5,7 @@ Currently the probe-rs project includes the following maintained & to be release
 - gdb-server (main release cycle)
 - probe-rs-cli (main release cycle)
 - probe-rs-cli-util (main release cycle)
-- probe-rs-t2rust (released optionally when changes are made)
+- probe-rs-target (released optionally when changes are made)
 - probe-rs-rtt (main release cycle)
 - rtt-target (separate rtt-target release cycle)
 - panic-rtt-target (separate rtt-target release cycle)
