--- probe-rs-t2rust/Cargo.toml
@@ -1,28 +1,0 @@
-[package]
-name = "probe-rs-t2rust"
-version = "0.7.0"
-authors = ["Dominik Boehi <dominik.boehi@gmail.com>"]
-edition = "2018"
-description = "A collection of ARM chips for the `probe-rs` crate."
-documentation = "https://docs.rs/probe-rs/"
-homepage = "https://github.com/probe-rs/probe-rs"
-repository = "https://github.com/probe-rs/probe-rs"
-readme = "../README.md"
-categories = ["embedded", "hardware-support", "development-tools::debugging"]
-keywords = ["embedded"]
-license = "MIT OR Apache-2.0"
-
-[[bin]]
-name = "probe-rs-t2rust"
-path = "src/bin.rs"
-
-[lib]
-name = "probe_rs_t2rust"
-path = "src/lib.rs"
-
-[dependencies]
-quote = "1.0.2"
-serde_yaml = "0.8.11"
-proc-macro2 = "1.0.4"
-base64 = "0.13.0"
-scroll = "0.10.1"
