+++ probe-rs-target/Cargo.toml
@@ -11,6 +11,13 @@ categories = ["embedded", "hardware-support", "development-tools::debugging"]
 keywords = ["embedded"]
 license = "MIT OR Apache-2.0"
 
+[features]
+# serde(skip_serializing_if) breaks with `bincode`, because it doesn't encode metadata (field names).
+# Enable this if you want to serialize to bincode. Do not enable this if you want to serialize to other formats (YAML, JSON...)
+# https://github.com/serde-rs/serde/issues/1732
+# https://github.com/bincode-org/bincode/issues/326
+bincode = []
+
 [dependencies]
 jep106 = "0.2.4"
 serde = { version = "1.0.104", features = ["derive"] }
