+++ probe-rs-target/src/chip.rs
@@ -14,7 +14,10 @@ pub struct Chip {
     pub name: Cow<'static, str>,
     /// The `PART` register of the chip.
     /// This value can be determined via the `cli info` command.
-    #[serde(skip_serializing_if = "Option::is_none")]
+    #[cfg_attr(
+        not(feature = "bincode"),
+        serde(skip_serializing_if = "Option::is_none")
+    )]
     pub part: Option<u16>,
     /// The memory regions available on the chip.
     pub memory_map: Cow<'static, [MemoryRegion]>,
