+++ probe-rs/Cargo.toml
@@ -44,14 +44,17 @@ anyhow = "1.0.31"
 bitvec = {version = "0.19.4" }
 funty = "=1.1.0" # Temporary fix for https://github.com/bitvecto-rs/bitvec/issues/105
 itm-decode = "0.1.5"
+bincode = "1.3.2"
 
 libftdi1-sys = { version = "1.0.0-alpha3", optional = true }
 static_assertions = "1.1.0"
 
-probe-rs-target  = { path = "../probe-rs-target", version ="0.1.0" }
+probe-rs-target  = { path = "../probe-rs-target", version = "0.1.0", features = ["bincode"] }
 
 [build-dependencies]
-probe-rs-t2rust  = { path = "../probe-rs-t2rust", version ="0.7.0" }
+probe-rs-target  = { path = "../probe-rs-target", version = "0.1.0", features = ["bincode"] }
+serde_yaml = "0.8.11"
+bincode = "1.3.2"
 
 [dev-dependencies]
 rand = "0.8.0"
