+++ probe-rs/targets/GD32VF1 Series.yaml
@@ -13,6 +13,6 @@ variants:
             start: 134217728
             end: 134348800
           is_boot_memory: true
-    flash_algorithms:
-flash_algorithms:
+    flash_algorithms: []
+flash_algorithms: {}
 core: riscv
